% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batson.app.pkg-package.R
\docType{package}
\name{batson.app.pkg-package}
\alias{batson.app.pkg}
\alias{batson.app.pkg-package}
\title{batson.app.pkg: Batson App}
\description{
This Shiny application estimates strike bias in the use of peremptory challenges to support Batson challenges.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sachinspandya/batson.app.pkg}
  \item Report bugs at \url{https://github.com/sachinspandya/batson.app.pkg/issues}
}

}
\author{
\strong{Maintainer}: Sachin Pandya \email{sachin.pandya@uconn.edu} (\href{https://orcid.org/0000-0001-7387-1307}{ORCID})

Authors:
\itemize{
  \item Xiaomeng Li
  \item Timothy Moore (\href{https://orcid.org/0000-0002-9576-0517}{ORCID})
}

}
\keyword{internal}
