% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batson.app.R
\docType{package}
\name{batson.app.pkg}
\alias{batson.app.pkg}
\title{batson.app.pkg: A Shiny app package for estimating bias in peremptory challenges}
\description{
This package provides one function to run the Shiny app:
run_app.
}
\section{Foo functions}{

The foo functions ...
}

